add --dump support and testing for missing --xclient-* options
Because I was working off of an old doc, I missed --xclient-delim, --xclient-before-starttls, --xclient-no-verify, xclient-destaddr, and --xclient-destport.  Add them to --dump output and testing as needed.

add util/gen-working-ref.pl to create ref.txt.  This is so the newest ref.txt can be used when looking something up during development.  base.pod could be used, but it's no fun to read raw pod