Fix issue with how .swaksrc is searched in paths of common env variables
In previous versions, a config file was set if the env variable was not empty.  In other words, regardless of whether $HOME/.swaksrc existed, if $HOME was not empty, the config file would be set to $HOME/.swaksrc. This had a real bug in terms of finding the correct file (if only $LOGDIR/.swaksrc existed, but $HOME was defined, config would be set to (the non-existent) $HOME/.swaksrc.

Second visible bug - if none of SWAKS_HOME/HOME/LOGDIR exist, DATA won't be checked

Basically, the thing basically worked, but only sort of, and kind of by accident.