use '-' instead of '--' to prefix bare config file and env var options
This is a bandaid for the issue where -foobarbaz if "--f oobarbaz" and --foobarbaz is invalid option "foobarbaz".

Found while writing test cases and trying to cause swaks to fail with "foobarbaz" in a config.  I could work around it for config by changing it to --foobarbaz, but there's no workaround for env vars, since they CAN'T be prefixed.

I hope to address this by figuring out how to require a space between the option and the arg.  I feel like it's turning off bundling, but for some reason I thought I looked into that and there was some reason I couldn't do it.