Add %NEWLINE% as a new --data token
A nice convenience for people who have issues understanding quoting on the command line, but added in the moment because I was having trouble with quoting \\n via expect in the test cases.  I originally got around it by just having a single line, but when I made the "no newlines means file" change, that then broke.