Improve error handling when the argument to --data looks like a file
If the arg to --data looks like a file but is not openable, error and exit instead of using it the file name as the raw data value

The motivation behind this is that previously, if a --data argument looked like a file but wasn't openable, swaks assumed it was the raw contents for DATA and sent it.  So, you might send an email with no (real) headers or body, with just text like "/path/to/file.txt".  It seems better to err on the side of caution and exit.  It seems very unlikely that a user would intentionally send an email w/ only one line.  If I'm wrong, I'll adjust the logic later.