Remove ambiguity from --attach* option processing
When processing --attach* options, they all end up in an array. We marked filenames so they were unambiguous, but determining what was a mime type and what was content was difficult and had at least one bug (setting an attachment with dir/file, when the file didn't exist, would treat it as a mime type rather than an attachment). Granted, we don't handle "I meant a file but it doesn't exist" very well either, but at least we handle it uniformly now.

This commit also lays necessary groundwork for --attach-body (wouldn't have been able to tell --attach from --attach-body otherwise)