major reworking of underlying option processing
Goals:
- abstract away from reading options from a raw data structure
- abstract away responsibility for common functionality

For instance, the "if option is provided but no argument, prompt the user for input" pattern was everywhere.  Why not define a function as that type and then abstract away the code that handles it?

Similarly, the --xclient-* and --attach-* family options had a very similar set of challenges (options which are different, and need to be processed differently, but they order they are specified in are meaningful to each other.  They were previously solved in two different and incompatible ways.

I also really want to do things like fix how file-reading code is handled uniquely wherever it's used.  The groundwork laid in this commit will allow me to do that.

process_args() as a whole could still use a ton of work, but this was the bulk change that moved option "types" into the option definition object instead of in code, and removed all direct reads on the specified-options structure.