fix handling of empty SWAKS_OPT_ environment variable options
Previously, swaks treated `SWAKS_OPT_foo` as `--foo ''`.  This meant that a lot of the built in "this option requires an argument" checks didn't work because an option was being provided, it was just empty.  Swaks will now treat this as a non-arg option, like `--foo`.

The docs say "Setting a variable to an empty value is the same as specifying it on the command line with no argument", so this is clearly a bug.