Fix character issue caused by localed date string (#25)
The issue was originally caused by 434f494abcc3558c73efc0e57a4338adeb402253, which was at attempt to fix timezone handling but mistakenly ended up localizing the text strings in the date, which was incorrect

Localizing the strings introduced characted issues for some locales that prevented the text from being rendered properly (specifically seen in cs_CZ.UTF-8) but only in some perls (did not happen in perl 5.18.4 (system) or 5.32.0 (brew) on macOS 10.15.6.  DID show up in perl 5.28.1 (system) on Debian 10.5.

If the fix for un-localizing the date had worked (3a95aa0498922a8a9591dda0da363f0e49b57086), the character rendering issue also would have been fixed.  However, the fix worked on macOS and failed on Debian, the same as above.

So, this commit is a fix for #17 that doesn't incorrectly introduce localized dates, which in turn caused #25.