Remove reliance on external tool for interactive tests
Just use the already-existing ability to capture output from and write input to an external script.  Allows us to enable interactive tests on windows

Pros: totally cross-platform, including addressing line ending issues we were experiencing on Windows
Cons: we're not actually responding directly to prompts.  This is slightly more fragile, but I think it's fine for a set of test tools.