fix unreliable verifcation of certificate hostname
A misunderstanding of how the openssl verify callback worked led to improperly failing hostname certification in some circumstances.  The solution is to just do "does this current cert satisfy or disqualify our needs" in the callback and then make the final decision after the connect(), when the callback has been called as many times as it's going to be

This bug was never released, it was introduced into development after the last prod release.