testing: fix/improve unmunge output from run-tests.pl
The goal of this action is to produce a copy/pastable version of the comand that was run for human debugging

bugfix: it did not properly preserve quoted arguments.  It turned "--pipe 'smtp-server.pl --arg 1 --foo 2'" into "--pipe smtp-server.pl --arg 1 --foo 2".  It now maintains the quoting
improvement: now adds backslashes and newlines after every option/arg pair to make it way easier to read